refactor: split archival scripts into several files
